+++ src/main/java/many/studio/web_backend/entity/Profissional.java
@@ -8,7 +8,7 @@ import java.time.LocalDateTime;
 public class Profissional {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Integer id;
+    private Long id;
 
     @Column(name = "nome")
     private String nome;
@@ -19,14 +19,14 @@ public class Profissional {
 
     public Profissional() {}
 
-    public Profissional(Integer id, String nome, String fotoUrl, LocalDateTime criadoEm, Usuario usuario) {
+    public Profissional(Long id, String nome, String fotoUrl, LocalDateTime criadoEm, Usuario usuario) {
         this.id = id;
         this.nome = nome;
         this.usuario = usuario;
     }
 
-    public Integer getId() { return id; }
-    public void setId(Integer id) { this.id = id; }
+    public Long getId() { return id; }
+    public void setId(Long id) { this.id = id; }
     public String getNome() { return nome; }
     public void setNome(String nome) { this.nome = nome; }
     public Usuario getUsuario() { return usuario; }
