+++ src/main/java/many/studio/web_backend/repository/ProfissionalRepository.java
@@ -3,5 +3,5 @@ package many.studio.web_backend.repository;
 import many.studio.web_backend.entity.Profissional;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface ProfissionalRepository extends JpaRepository<Profissional, Integer> {
+public interface ProfissionalRepository extends JpaRepository<Profissional, Long> {
 }
