+++ src/main/java/many/studio/web_backend/service/UsuarioService.java
@@ -100,7 +100,7 @@ public class UsuarioService {
                 .toList();
     }
 
-    public void atualizar(Integer id, UsuarioAtualizarDto dto) {
+    public void atualizar(Long id, UsuarioAtualizarDto dto) {
 
         Usuario usuarioExistente = usuarioRepository.findById(id)
                 .orElseThrow(() -> new EntityNotFoundException("Usuário não encontrado"));
@@ -119,7 +119,7 @@ public class UsuarioService {
         usuarioRepository.save(usuarioExistente);
     }
 
-    public void removerPorId(Integer id) {
+    public void removerPorId(Long id) {
         if (!usuarioRepository.existsById(id)) {
             throw new EntityNotFoundException("Usuário não encontrado");
         }
